bot/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.1.0
requests==2.32.3
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
